Move ComputeStatistics into pipeline_utils and setup v2
