+++ src/SOPRANO/dnds.py
@@ -1,12 +1,8 @@
 import numpy as np
 import pandas as pd
 
-from SOPRANO.objects import AnalysisPaths, AuxiliaryPaths, Parameters
-from SOPRANO.pipeline_utils import (
-    MissingDataError,
-    _PipelineComponent,
-    is_empty,
-)
+from SOPRANO.objects import AnalysisPaths, AuxiliaryPaths
+from SOPRANO.pipeline_utils import is_empty
 from SOPRANO.sh_utils import subprocess_pipes
 
 
@@ -45,26 +41,6 @@ def _intersect_introns(paths: AnalysisPaths, aux_files: AuxiliaryPaths):
         ["awk", r'{print $4"\t"$8/($6+1)"\t"$8"\t"$6}'],
         output_path=paths.intron_rate,
     )
-
-
-class ComputeStatistics(_PipelineComponent):
-    @staticmethod
-    def check_ready(params: Parameters):
-        paths = (
-            params.data_epitopes,
-            params.epitope_nans,
-            params.intra_epitope_nans,
-            params.intron_rate,
-        )
-
-        for path in paths:
-            if not path.exists():
-                raise MissingDataError(path)
-
-    @staticmethod
-    def apply(params: Parameters):
-        ComputeStatistics.check_ready(params)
-        _compute_coverage(params)
 
 
 def _preprocess_dfs(paths: AnalysisPaths):
